requirements.txt
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
datasets==3.2.0
dill==0.3.8
Django==5.0.7
django-cors-headers==4.6.0
django-grappelli==4.0.1
djangorestframework==3.15.2
drf-yasg==1.21.7
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.27.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
propcache==0.2.1
psutil==6.1.1
psycopg2==2.9.9
pyarrow==18.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sqlparse==0.5.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
